OLD:
```
Benchmark_Ctx_BodyParser_JSON
Benchmark_Ctx_BodyParser_JSON-4                  1452151               817 ns/op             368 B/op          7 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1461086               821 ns/op             368 B/op          7 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1458460               819 ns/op             368 B/op          7 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1430881               825 ns/op             368 B/op          7 allocs/op
Benchmark_Ctx_BodyParser_XML
Benchmark_Ctx_BodyParser_XML-4                    341536              2947 ns/op            1312 B/op         26 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    404234              2923 ns/op            1312 B/op         26 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    366728              2958 ns/op            1312 B/op         26 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    369998              2938 ns/op            1312 B/op         26 allocs/op
Benchmark_Ctx_BodyParser_Form
Benchmark_Ctx_BodyParser_Form-4                   519937              2133 ns/op             888 B/op         24 allocs/op
Benchmark_Ctx_BodyParser_Form-4                   521875              2127 ns/op             888 B/op         24 allocs/op
Benchmark_Ctx_BodyParser_Form-4                   540964              2143 ns/op             888 B/op         24 allocs/op
Benchmark_Ctx_BodyParser_Form-4                   511980              2143 ns/op             888 B/op         24 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm
Benchmark_Ctx_BodyParser_MultipartForm-4          525433              2023 ns/op             872 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4          575828              2055 ns/op             872 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4          539637              2035 ns/op             872 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4          535551              2053 ns/op             872 B/op         23 allocs/op
```

NEW:
```
Benchmark_Ctx_BodyParser_JSON
Benchmark_Ctx_BodyParser_JSON-4                  1793932               659 ns/op             208 B/op          4 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1809279               660 ns/op             208 B/op          4 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1774881               660 ns/op             208 B/op          4 allocs/op
Benchmark_Ctx_BodyParser_JSON-4                  1788525               661 ns/op             208 B/op          4 allocs/op
Benchmark_Ctx_BodyParser_XML
Benchmark_Ctx_BodyParser_XML-4                    409476              2811 ns/op            1152 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    400009              2797 ns/op            1152 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    410803              2787 ns/op            1152 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_XML-4                    399465              2794 ns/op            1152 B/op         23 allocs/op
Benchmark_Ctx_BodyParser_Form
Benchmark_Ctx_BodyParser_Form-4                   991834              1189 ns/op             312 B/op         12 allocs/op
Benchmark_Ctx_BodyParser_Form-4                  1000000              1195 ns/op             312 B/op         12 allocs/op
Benchmark_Ctx_BodyParser_Form-4                   981740              1197 ns/op             312 B/op         12 allocs/op
Benchmark_Ctx_BodyParser_Form-4                   980721              1188 ns/op             312 B/op         12 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm
Benchmark_Ctx_BodyParser_MultipartForm-4         1000000              1115 ns/op             296 B/op         11 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4         1000000              1119 ns/op             296 B/op         11 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4         1000000              1119 ns/op             296 B/op         11 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm-4         1000000              1120 ns/op             296 B/op         11 allocs/op
```